+++ NEWS.md
@@ -1,3 +1,10 @@
+# madplan 0.2.0
+
+Made the following changes:
+
+* Including a download button for windows allowing to download the food plan
+* Added `pullToRefresh = TRUE` in the options for `shinyMobile::f7Page`
+
 # madplan 0.1.0
 
 * End of line fix such that data with line breaks can be saved within the same vector while still having neat rendering